requirements.txt
mss>=10.1.0
opencv-python>=4.12.0
pytesseract>=0.3.13
sounddevice>=0.5.3
vosk>=0.3.45
numpy>=2.2.6
pyinstaller>=6.16.0
pillow>=12.0.0
scipy>=1.16.2
